Navbar accounts, home page blocks
Added Meteor accounts to the navbar which includes login with email, create account, and reset password all in a modal box. Other links in the navbar were updated so that their styles match that of the accounts component. Dashboard and survey replaced items displayed on the home page and were configured with the accounts packaging to ensure links to either will only be visible if the user is logged in. Inspirational headers were added to blank spaces on the homepage and placeholder titles were added to the subsections that should eventually contain information about BirthdayCrate in the top, and about the development team in the bottom. Up next in my sprint will be starting up a dashboard for users which will allow them to view their orders. Next the survey page will include Jiayu's survey through Gabi's selection page. Content will be added to the home page informational sections if nothing is submitted by the intended author.